DEFRESTART: Document restarts and define convenience function to invoke.
E.g. (defrestart foo "Restart to foobar, establish by `BAZ'") will
define a function foo that can be invoked to invoke the restart (so,
using (foobar) instead of (invoke-restart foobar).

At the same time the given docstring will be attached to the function
allowing restart documentation to be written with the corresponding
symbol and be easily referenced from other doc-strings.

E.g.

  (defun baz ()
     "
     Some documentation on baz.
     ...

     Restarts: `FOOBAR', `ABORT-BAZ'.
   "
   ...)