WITH-GENSYMS: Generating symbols for macros with names that contain their placeholder names.
   (with-gensyms (foo bar)
      BODY)

Will generate new uninterned symbols and bind them to FOO and BAR. The
difference to "usual" with-gensyms implementation is, that the
generated symbols will contain the names of the variables to which
they're bound, e.g. in this case #:G.FOO.131 and #:G.BAR.132, which
helps a lot when looking at the variables in the debugger and costs
practically nothing.